extracting prometheus calls from passive health checker
